+++ src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { AuthenticationModule } from './authentication/authentication.module';
 import { MysqlDatabaseModule } from './database/mysql/database.module';
-import { RequestsModule } from './models/requests/requests.module';
+import { RequestsModule } from './modules/requests/requests.module';
 import { HttpExceptionFilter } from './common/exceptions/http-error.filter';
 import { APP_FILTER } from '@nestjs/core';
