+++ src/modules/requests/requests.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { RequestsService } from './requests.service';
+import { RequestsService } from './services/requests.service';
 import { RequestsController } from './requests.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Request } from './entities/request.entity';
-import { RequestsRepository } from './requests.repository';
+import { RequestsRepository } from './respositories/requests.repository';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Request])],
